Google maps successfully plotting location, name, and address of venue along with event name
